Created the state waiting_remote_coordinator and started a timer to trigger the coordinator_timeout.
The waiting_remote_coordinator state is simply waiting for the {ack, Node, now_executing} message, which is sent using erlang:send/2, so there is no need to have state functions for that state.
If anything but the ack message arrives in that state things are really broken and should fail instantly.